moved game_manager to new cpp
wrote headers
