+++ usr/Scores_list_10.txt
@@ -1,14 +1,14 @@
-YAY 90
+YAY 120
 
 YAY 90
 
-YAY 80
+YAY 90
 
 YAY 80
 
 YAY 80
 
-YAY 20
+YAY 80
 
 YAY 20
 
@@ -16,7 +16,7 @@ YAY 20
 
 YAY 20
 
--- 10
+YAY 20
 
 -- 10
 
@@ -25,6 +25,8 @@ YAY 20
 YAY 10
 
 YAY 10
+
+-- 10
 
 -- 10
